Don't unlink() the file when hiding.
When we want to hide a file, we currently rename() the old file to hide
it.  But there is a period of time in which the new filename doesn't
exist, which breaks atomicity.  So call link() instead of rename() in
this case, so that the old file continues to exist.