Kludge: Don't unhash nodes, to prevent race.
Sometimes, when we're unhashing a node, we accidentally unhash a node
that has been queued, which results in overwritten files failing to
exist.  The correct solution would be to detect that the node has been
queued, and don't unhash it then, but that seems to be a fairly major
undertaking.